requirements.txt
streamlit==1.51.0
google-generativeai==0.8.5
Pygments==2.19.2
requests==2.32.3
openai==1.12.0
